Increase char limit for direct schedule comments
